--- packages/behandling-opplaeringspenger/src/panelDefinisjoner/faktaPaneler/MedisinskVilkarFaktaPanelDef2.tsx
@@ -1,32 +1,0 @@
-import React from 'react';
-
-import { faktaPanelCodes } from '@k9-sak-web/konstanter';
-import aksjonspunktCodes from '@fpsak-frontend/kodeverk/src/aksjonspunktCodes';
-import { FaktaPanelDef } from '@k9-sak-web/behandling-felles';
-import { Fagsak, Behandling } from '@k9-sak-web/types';
-
-import MedisinskVilkår from '../../components/MedisinskVilkår';
-
-class MedisinskVilkarFaktaPanelDef2 extends FaktaPanelDef {
-  getUrlKode = () => faktaPanelCodes.MEDISINSKVILKAAR_V2;
-
-  getTekstKode = () => 'MedisinskVilkarPanel.MedisinskVilkar';
-
-  getAksjonspunktKoder = () => [aksjonspunktCodes.MEDISINSK_VILKAAR];
-
-  getEndepunkter = () => [];
-
-  getKomponent = props => <MedisinskVilkår {...props} />;
-
-  getData = ({ fagsak, behandling }: { fagsak: Fagsak; behandling: Behandling }) => ({
-    fagsakYtelseType: fagsak.sakstype,
-    behandlingType: behandling.type.kode,
-  });
-
-  getOverstyrVisningAvKomponent = ({ behandling }: { fagsak: Fagsak; behandling: Behandling }) => {
-    const søknadsfristErIkkeUnderVurdering = behandling.stegTilstand?.stegType?.kode !== 'VURDER_SØKNADSFRIST';
-    return søknadsfristErIkkeUnderVurdering;
-  };
-}
-
-export default MedisinskVilkarFaktaPanelDef2;
