+++ packages/behandling-opplaeringspenger/src/panelDefinisjoner/prosessStegOpplaeringspengerPanelDefinisjoner.tsx
@@ -3,6 +3,7 @@ import FortsattMedlemskapProsessStegPanelDef from './prosessStegPaneler/Fortsatt
 import InngangsvilkarFortsProsessStegPanelDef from './prosessStegPaneler/InngangsvilkarFortsProsessStegPanelDef';
 import InngangsvilkarProsessStegPanelDef from './prosessStegPaneler/InngangsvilkarProsessStegPanelDef';
 import MedisinskVilkarProsessStegPanelDef from './prosessStegPaneler/MedisinskVilkarProsessStegPanelDef';
+import OpplaeringProsessStegPanelDef from './prosessStegPaneler/OpplaeringProsessStegPanelDef';
 import SimuleringProsessStegPanelDef from './prosessStegPaneler/SimuleringProsessStegPanelDef';
 import TilkjentYtelseProsessStegPanelDef from './prosessStegPaneler/TilkjentYtelseProsessStegPanelDef';
 import UttakProsessStegPanelDef from './prosessStegPaneler/UttakProsessStegPanelDef';
@@ -12,7 +13,7 @@ import VedtakProsessStegPanelDef from './prosessStegPaneler/VedtakProsessStegPan
 const prosessStegPanelDefinisjoner = [
   new VarselProsessStegPanelDef(),
   new InngangsvilkarProsessStegPanelDef(),
-  new MedisinskVilkarProsessStegPanelDef(),
+  new OpplaeringProsessStegPanelDef(),
   new InngangsvilkarFortsProsessStegPanelDef(),
   new FortsattMedlemskapProsessStegPanelDef(),
   new BeregningsgrunnlagProsessStegPanelDef(),
