+++ packages/behandling-pleiepenger-sluttfase/i18n/nb_NO.json
@@ -9,6 +9,7 @@
   "MainSideMenu.Heading": "Saksopplysninger",
   "OpptjeningInfoPanel.KontrollerFaktaForOpptjening": "Opptjening",
   "InntektsmeldingInfoPanel.Title": "Inntektsmelding",
+  "Behandlingspunkt.SykdomOgOpplaering": "Sykdom og opplæring",
   "Behandlingspunkt.LivetsSluttfase": "Livets sluttfase",
   "Behandlingspunkt.Inngangsvilkar": "Inngangsvilkår",
   "Behandlingspunkt.InngangsvilkarForts": "Inngangsvilkår Fortsettelse",
