+++ packages/v2/gui/src/shared/labelled-content/LabelledContent.tsx
@@ -13,7 +13,8 @@ export interface LabelledContentProps {
 
 export const LabelledContent = ({ label, content, labelTag, indentContent, size = 'small' }: LabelledContentProps) => {
   const cl = classNames('mt-2', {
-    'border-solid border-t-0 border-r-0 border-b-0 border-l-4 border-[#c9c9c9] py-[3px] pl-[13px]': indentContent,
+    'border-solid border-t-0 border-r-0 border-b-0 border-l-4 border-[#c9c9c9] py-[3px] pl-[13px]':
+      content && indentContent,
   });
   return (
     <div>
