+++ packages/v2/gui/src/shared/vurderingsperiode-navigasjon/VurderingsperiodeNavigasjon.tsx
@@ -13,6 +13,7 @@ import {
   ReisetidPeriodeVurderingDtoResultat as reisetidEnumObject,
 } from '@k9-sak-web/backend/k9sak/generated';
 import styles from './periodeRad.module.css';
+import { element } from '../basicList/basicList.module.css';
 
 export type ResultatType =
   | InstitusjonVurderingDtoResultat
@@ -74,7 +75,7 @@ const Vurderingsnavigasjon = <T extends Vurderingselement = Vurderingselement>({
       onPeriodeClick(allePerioder[index]);
     }
   };
-
+  console.log(allePerioder)
   return (
     <Box className="min-w-[400px]">
       <Heading size="xsmall" className="ml-[15px] mt-[21px] mb-[24px]">
@@ -82,7 +83,6 @@ const Vurderingsnavigasjon = <T extends Vurderingselement = Vurderingselement>({
       </Heading>
 
       {allePerioder.length === 0 && <div className="ml-[15px] mt-[15px] mb-5">Ingen vurderinger å vise</div>}
-
       {allePerioder.length > 0 && (
         <>
           <div className="flex w-[120px]">
